requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-filter==2.2.0
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
neobolt==1.7.15
neotime==1.7.4
oauthlib==3.1.0
packaging==19.2
prompt-toolkit==2.0.10
py2neo==4.3.0
Pygments==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
